imagem do paciente implementada

+++ POV administrador/views/consultas-adm/src/app/forms/edita-vg/editar-vg/editar-vg.component.ts
@@ -16,12 +16,13 @@ import { Location } from '@angular/common';
   templateUrl: './editar-vg.component.html',
   styleUrl: './editar-vg.component.css'
 })
-export class EditarVgComponent implements OnInit{
+export class EditarVgComponent implements OnInit {
   //id pegado por url
   id!: string;
   formulario!: FormGroup;
   consultas: Consulta[] = []; //lista consulta
   consulta!: Consulta; //consulta editada
+  edit: boolean = false;
   //inputs
   @ViewChild('tipo') tipo!: ElementRef<HTMLInputElement>;
   @ViewChild('preco') preco!: ElementRef<HTMLInputElement>;
@@ -35,13 +36,14 @@ export class EditarVgComponent implements OnInit{
 
   ngOnInit(): void {
     this.route.params.subscribe(params => {
-      this.id = params['id']; 
-      this.consultaService.getConsultas().subscribe( dado => {
+      this.id = params['id'];
+      this.consultaService.getConsultas().subscribe(dado => {
         this.consultas = dado;
         this.consulta = this.buscarConsulta(this.id)!;
-        this.tipo.nativeElement.value = this.consulta.tipo_consulta;      
+        this.tipo.nativeElement.value = this.consulta.tipo_consulta;
         this.data.nativeElement.value = this.consulta.data_consulta;
         this.preco.nativeElement.value = this.consulta.preco.toString();
+        this.verificarConsulta(this.consulta.data_consulta);
       });
     });
   }
@@ -55,20 +57,28 @@ export class EditarVgComponent implements OnInit{
     return null;
   }
 
+  verificarConsulta(data: string) {
+    const dataConsulta = new Date(data); // Converte para objeto Date
+    const agora = new Date(); // Data e hora atual
+    if (dataConsulta < agora) {
+      this.edit = true;
+    }
+  }
+
   //metodo de atualizar consulta
-  salvarConsulta(){
+  salvarConsulta() {
     //nova consulta com os dados novos
     const novaConsulta = {
-      id : this.consulta.id,
+      id: this.consulta.id,
       tipo_consulta: this.tipo.nativeElement.value,
-      preco:  Number(this.preco.nativeElement.value),
+      preco: Number(this.preco.nativeElement.value),
       data_consulta: this.data.nativeElement.value,
       cpf_paciente: '',
       id_clinica: this.consulta.id_clinica,
       relatorio: this.consulta.relatorio
     }
     //verificando se é para apagar o cliente
-    if(this.apagar.nativeElement.value == 'nao'){
+    if (this.apagar.nativeElement.value == 'nao') {
       novaConsulta.cpf_paciente = this.consulta.cpf_paciente;
     }
     //mandando pro servidor a nova consulta a ser substituida
@@ -78,12 +88,15 @@ export class EditarVgComponent implements OnInit{
     toastElement.show();
   }
 
-  apagarConsulta(){
-    //apagar a consulta do bd
-    this.consultaService.deleteConsulta(this.consulta).subscribe(() => {
-      alert('Excluído Consulta com Sucesso, Você será redirecionado!');
-      this.location.back(); // Volta para a página anterior
-    });
+  apagarConsulta() {
+    var resposta = confirm('Você deseja mesmo apaga a consulta?');
+    if (resposta) {
+      //apagar a consulta do bd
+      this.consultaService.deleteConsulta(this.consulta).subscribe(() => {
+        alert('Excluído Consulta com Sucesso, Você será redirecionado!');
+        this.location.back(); // Volta para a página anterior
+      });
+    }
   }
 
 }
